requirements.txt
pyrogram>=2.0.0
tgcrypto>=1.2.0
vkmax>=1.0.2
pyyaml>=6.0
aiohttp>=3.9.0
